requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5

pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
seaborn==0.13.2
